+++ test/expression/node/Node.test.js
@@ -1,0 +1,32 @@
+// test Node
+var assert = require('assert'),
+    approx = require('../../../tools/approx'),
+    math = require('../../../index')();
+
+describe('Node', function() {
+
+  it ('should create a Node', function () {
+    // TODO
+  });
+
+  it ('should evaluate a Node', function () {
+    // TODO
+  });
+
+  it ('should compile a Node', function () {
+    // TODO
+  });
+
+  it ('should find a Node', function () {
+    // TODO
+  });
+
+  it ('should match a Node', function () {
+    // TODO
+  });
+
+  it ('should stringify a Node', function () {
+    // TODO
+  });
+
+});