+++ test/expression/node/UpdateNode.test.js
@@ -1,0 +1,32 @@
+// test UpdateNode
+var assert = require('assert'),
+    approx = require('../../../tools/approx'),
+    math = require('../../../index')();
+
+describe('UpdateNode', function() {
+
+  it ('should create a UpdateNode', function () {
+    // TODO
+  });
+
+  it ('should evaluate a UpdateNode', function () {
+    // TODO
+  });
+
+  it ('should compile a UpdateNode', function () {
+    // TODO
+  });
+
+  it ('should find a UpdateNode', function () {
+    // TODO
+  });
+
+  it ('should match a UpdateNode', function () {
+    // TODO
+  });
+
+  it ('should stringify a UpdateNode', function () {
+    // TODO
+  });
+
+});