+++ test/expression/node/handlers.test.js
@@ -1,0 +1,12 @@
+// test handlers
+var assert = require('assert'),
+    approx = require('../../../tools/approx'),
+    math = require('../../../index')();
+
+describe('handlers', function() {
+
+  it ('should create custom Node handlers', function () {
+    // TODO
+  });
+
+});